test: separate test into projects; add teardown step to API mock
